pyproject.toml
[tool.poetry]
name = "openai-proxy"
version = "0.1.0"
description = ""
authors = ["celeste <parafactual@gmail.com>"]
readme = "README.md"
packages = [{include = "openai_proxy"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
openai = "^0.27.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
